feat(devcontainer)!: switch from `poetry` to `uv` to handle python dependencies as well as python interpreter versions'
The most important change is the switch from `poetry` to `uv` (which is the reason for the breaking
change). However, the stage was almost completely rewritten. All shell steps including pipes were
rewritten to use no pipes at all in order to not rely on a shell change with pipefail. The
corresponding line was removed. Because `uv` is now used to handle python dependencies **and**
interpreter versions the `PYENV_VERSION` and `PYTHON_VERSION` build arguments were removed. Instead
of relying on `uname` to detect the machine architecture the special build argument `TARGETPLATFORM`
was introduced. This is handled by BuildKit automatically. Additionally, `cocogitto` for release
management of projects was added.